menu

+++ states/impl/running_state.py
@@ -26,7 +26,7 @@ from states.workers import obstacle_generator_worker
 class RunningState(GameState):
     def __init__(self, context):
         self.is_game_over = False
-        self.difficulty_level = Difficulty1()
+        self.difficulty_class_level = Difficulty1()
         self.active_obstacles = deque()
         self.obstacles_to_render = deque()
         self.obstacles_per_frame = 1
@@ -42,9 +42,8 @@ class RunningState(GameState):
             ObstacleLongCube,
             ObstacleWoodenSign,
             ObstacleTrain
-        ], max_size_per_type=50)
+        ], max_size_per_type=15)
         self.difficulty_manager = DifficultyManager()
-        self.start()
         self.difficulty_level = multiprocessing.Value('i', 1)
         self.go = multiprocessing.Value('b', True)
         self.player_z = multiprocessing.Value('d', 0.0)
@@ -54,6 +53,9 @@ class RunningState(GameState):
             args=(self.obstacle_queue, self.player_z, self.go, self.difficulty_level)
         )
         self.obstacle_process.start()
+        self.start()
+        self.create_paused_panel()
+
 
     def on_exit(self):
         for obstacle in self.active_obstacles:
@@ -62,9 +64,10 @@ class RunningState(GameState):
         self.active_obstacles.clear()
         print('exiting state')
         self.go.value = False
-        self.obstacle_process.terminate()
-        self.obstacle_process.join()
-        del self.obstacle_process
+        if hasattr(self, "obstacle_process"):
+            self.obstacle_process.terminate()
+            self.obstacle_process.join()
+            del self.obstacle_process
 
     def handle_input(self):
         self.context.player.reset()
@@ -80,6 +83,8 @@ class RunningState(GameState):
             self.context.camera.camera.y = self.context.camera.camera.y - 0.3
         if held_keys['l']:
             self.set_difficulty(1)
+        if held_keys['escape']:
+            self.__toggle_paused()
         if held_keys['h']:
             self.set_difficulty(2)
         if held_keys['space'] and not self.context.player.is_jumping:
@@ -109,12 +114,33 @@ class RunningState(GameState):
         self.__render_obstacles_from_queue()
         self.__cleanup_obstacles()
 
+    def create_paused_panel(self):
+        self.pause_panel = WindowPanel(
+            title='Resume',
+            content=(
+                Button('Resume', color=color.gray, on_click=self.__toggle_paused),
+                Button('Main Menu', color=color.gray, on_click=self.__transition_to_main_menu),
+                Button('Quit', color=color.red, on_click=application.quit)
+            ),
+            position=(0, 0.25),
+            enabled=False
+        )
+
     def set_difficulty(self, level, **kwargs):
         self.difficulty_level.value = level
         self.difficulty_manager.set_player_settings(level, self.context.player)
 
+    def __toggle_paused(self):
+        application.paused = not application.paused
+        self.pause_panel.enabled = application.paused
+
+    def __transition_to_main_menu(self):
+        self.__toggle_paused()
+        self.on_exit()
+        self.context.transition_to('main_menu')
+
     def __initialize_obstacles(self):
-        obstacles = self.difficulty_level.initialize_obstacles()
+        obstacles = self.difficulty_class_level.initialize_obstacles()
         for obstacle_type in obstacles:
             obstacle = self.obstacle_pool.acquire(obstacle_type.obstacle, position_z=obstacle_type.position_z,
                                                   difficulty=obstacle_type.difficulty, lane=obstacle_type.lane, metadata=obstacle_type.entity_metadata)
@@ -127,12 +153,17 @@ class RunningState(GameState):
         obstacles_to_remove = []
         for obstacle in list(self.active_obstacles):
             if obstacle.z < player_z - self.cleanup_threshold:
+                print(f'remove obst {obstacle.z}, player {player_z}')
                 obstacles_to_remove.append(obstacle)
 
         for obstacle in obstacles_to_remove:
+            print(self.active_obstacles)
             self.active_obstacles.remove(obstacle)
-            obstacle.delete()
-            destroy(obstacle)
+            print(f'after removing {self.active_obstacles}')
+            # obstacle.delete()
+            # destroy(obstacle)
+            self.obstacle_pool.release(obstacle)
+        obstacles_to_remove.clear()
 
     def __render_obstacles_from_queue(self):
         for _ in range(self.obstacles_per_frame):
@@ -146,6 +177,7 @@ class RunningState(GameState):
                 position_z=obstacle_type.position_z,
                 difficulty=obstacle_type.difficulty,
                 lane=obstacle_type.lane,
-                metadata = obstacle_type.entity_metadata
+                metadata=obstacle_type.entity_metadata
             )
             self.active_obstacles.append(obstacle)
+            print(f'after adding {self.active_obstacles}')
